+++ service/local_lookup/character.py
